Revert "fix ui chat"
This reverts commit 3d5d8349bf58c3243df3366aeee4a5d03e5a9ffa.
Revert "add processbar"

This reverts commit 94e43eaef48f9fe5ab47cebf80ec9f69405643c1.

+++ app/src/main/java/com/example/dattingapp/utils/SharedPreference.java
@@ -101,7 +101,7 @@ public class SharedPreference {
         defaultFilter.maxAge = 150;
         defaultFilter.minAge =0;
         defaultFilter.gender = "Both";
-        defaultFilter.distance = Double.MAX_VALUE;
+        defaultFilter.distance = Integer.MAX_VALUE;
         if(filterJson == ""){
             return defaultFilter;
         }
